Fix bugs with calculator
The calculator now only evaluates a single pair of numbers at a
time. Entering a number, then operator, then another number, then
another operator will cause the calculator to display the result
of the first operation. Entering another number followed by = or
another operator will show the result of the second operation and
so on.

Numbers are now rounded to up to 8 decimal places by adding
+(operate(num1, op, num2).toFixed(8).

Pressing = before completing a full operation will now result in
the calculator displaying the first number or 0 depending on
whether the user entered a number first or not. If the user enters
a number and an operator, then clicks =, the calculator will assume
the second number is 0 and display the result. This is a problem if
the user uses the * or especially the / operator. This will be worked
out next.